WIP updating to latest manifold release
Next hurdle: resolve the requirement for dllclipper2.so. Currently only
the static archive .a is provided by the manifold build, which is a
problem for ctypes (as I'm using it with cstubs static building).
Switching to dynamic with foreign would be a much larger change I think,
so sticking with the current way would be ideal.